requirements.txt
# Core ML/AI Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scikit-learn>=1.3.0

# System Monitoring
psutil>=5.9.0

# Data Handling
pyyaml>=6.0
toml>=0.10.2

# Async Operations
aiofiles>=23.0.0

# Additional utilities
python-dateutil>=2.8.2
tqdm>=4.65.0

# File monitoring
watchdog>=3.0.0

# HTTP requests (for Ollama)
requests>=2.31.0